[feat] add lobby to alias through api / omaservice. changed LobbyId to long in model, as it should be. changed datservice bool return expressions.

+++ Models/Lobby.cs
@@ -3,7 +3,7 @@ namespace OMA.Models;
 public record Lobby
 {
     public int Id { get; set; }
-    public int LobbyId { get; set; }
+    public long LobbyId { get; set; }
     public int BestOf { get; set; } = 0;
     public int Warmups { get; set; } = 0;
     public List<Alias> Aliases { get; set; } = [];
